Added Building Manager Script
Added Building Manager a Script:  This script will manage the placement of placing buildings within the game. Currently I added a function to get the mouse position and convert it to world space for placing buildings within the world with percision. Also Added a place holder sprite for the wood gathering building that will later be updated to better match the theme of the game.